chore: silence two more clippy errors gated by metal,accelerate features
The earlier sweep ran without `--features metal,accelerate`, so the CI
build (which uses that exact feature set) still tripped on two finds:

- `quant.rs:422` unnecessary `as usize` cast on an already-usize index
- `generate.rs:826` explicit `.into_iter()` inside `zip(...)` which
  already accepts any `IntoIterator`

`cargo clippy --workspace --all-targets --features metal,accelerate
-- -D warnings` is now clean.

+++ src/lib/mlxcel-core/src/cache/turbo/quant.rs
@@ -419,7 +419,7 @@ fn quantize_into_packed(
         let idx_off = tok * coords_per_token;
         let mut sum_sq = 0.0_f32;
         for d_i in 0..coords_per_token {
-            let c = centroids[indices[idx_off + d_i] as usize];
+            let c = centroids[indices[idx_off + d_i]];
             sum_sq += c * c;
         }
         let y_hat_norm = sum_sq.sqrt();
